requirements.txt
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.6
contourpy==1.0.7
crashtest==0.3.1
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
diffusers==0.10.2
distlib==0.3.5
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
Faker==18.7.0
favicon==0.7.0
filelock==3.7.1
Flask==2.2.3
flask-ngrok==0.0.25
fonttools==4.39.4
frozenlist==1.3.3
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
greenlet==1.1.3
gspread==5.4.0
htbuilder==0.6.1
html5lib==1.1
httplib2==0.20.4
hugchat==0.0.5.6
huggingface-hub==0.13.4
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.12.0
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.4
jupyter_core==4.11.2
keyring==23.7.0
kiwisolver==1.4.4
langchain==0.0.162
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.0.10
packaging==20.9
pandas==1.4.3
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.14
poetry-core==1.0.8
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pydantic==1.10.7
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.13.0
pylev==1.4.0
pymdown-extensions==9.11
Pympler==1.0.1
PyMsgBox==1.0.9
pyngrok==4.1.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
PyRect==0.1.4
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.2
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
regex==2023.3.23
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==13.3.5
rsa==4.8
scipy==1.10.1
shellingham==1.4.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
st-annotated-text==4.0.0
stack-data==0.6.0
streamlit==1.22.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-chat==0.0.2.2
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.11.1
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.11.1
toolz==0.12.0
torch==2.0.0
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
virtualenv==20.15.1
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
yarl==1.9.1
youtube-dl==2021.12.17
zipp==3.8.1
